requirements.txt
numpy
pandas
streamlit
openpyxl
azure.storage.blob
azure_core